Structures package for publication - `1.0.0-beta` - Moves ./lib/components/ -> ./lib/ - Adds `make publish` and simplifies Makefile - Adds .npmignore - Removes webpack prod config in favor of only using webpack for demo construction
