+++ .npmignore
@@ -1,0 +1,4 @@
+.babelrc
+demo
+Makefile
+webpack.config.js
