+++ Makefile
@@ -1,7 +1,6 @@
 BIN = ./node_modules/.bin
 WEBPACK = $(BIN)/webpack
 SERVER = $(BIN)/webpack-dev-server --inline --hot --port 8040
-SRC = $(shell find . -type f -name '*.js' ! -path './build/*' -o -name '*.css' ! -path './build/*')
 
 server: build
 ifdef RECURLY_JS_CERT
@@ -10,16 +9,15 @@ else
 	@$(SERVER) --https
 endif
 
-build: demo
+build: build/demo.js
+build/demo.js: demo lib node_modules
+	@$(WEBPACK) --display-reasons --display-chunks
 
 test:
 	@npm test
 
-demo: lib
-	@$(WEBPACK) --display-reasons --display-chunks
-
-lib: $(SRC) node_modules
-	@$(WEBPACK) --display-reasons --display-chunks --config webpack.prod.config.js
+publish: lib clean node_modules
+	@npm publish --access public
 
 node_modules: package.json
 	@npm install
@@ -27,4 +25,4 @@ node_modules: package.json
 clean:
 	@rm -rf node_modules build
 
-.PHONY: test server clean
+.PHONY: clean publish server test
