+++ lib/index.js
@@ -1,12 +1,12 @@
-import RecurlyProvider from './components/provider';
-import useRecurly from './components/use-recurly';
-import Elements from './components/elements';
-import CardElement from './components/element/card';
-import CardNumberElement from './components/element/card-number';
-import CardMonthElement from './components/element/card-month';
-import CardYearElement from './components/element/card-year';
-import CardCvvElement from './components/element/card-cvv';
-import ThreeDSecureAction from './components/three-d-secure-action'
+import RecurlyProvider from './provider';
+import useRecurly from './use-recurly';
+import Elements from './elements';
+import CardElement from './element/card';
+import CardNumberElement from './element/card-number';
+import CardMonthElement from './element/card-month';
+import CardYearElement from './element/card-year';
+import CardCvvElement from './element/card-cvv';
+import ThreeDSecureAction from './three-d-secure-action'
 
 export {
   RecurlyProvider,
